pyproject.toml
[project]
name = "motion_demo"
version = "0.0.1"
description = "Motion Capture Demo"
authors = [{name = "Lance Borden", email = "borden.108@osu.edu"}]
readme = "README.md"
requires-python = "==3.11.13"

dependencies = [
    "mediapipe-numpy2>=0.10.21",
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0",
    "pyqt5>=5.15.11",
]

[dependency-groups]
dev = [
    "ruff>=0.5.3",
    "mypy>=1.10.0",
    "safety>=3.2.1",
    "pre-commit>=4.2.0",
    "pre-commit-hooks>=5.0.0",
]

[tool.ruff]
line-length = 140
show-fixes = true
exclude = ["yolov5"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
extend-select = [
    "I",     # isort
    "N",     # naming conventions
    "UP",    # pyupgrade -- catches obsolete code patterns
    "ANN",   # type annotations
    "ASYNC", # async code checks
    "B",     # flake8-bugbear -- catches bug-prone usage
    "C4",    # list comprehensions
    "DTZ",   # datetime mistakes -- also please use pendulum instead of datetime
    "INP",   # checks for presence of __init__.py
    "TID",   # tidy imports
    "PTH",   # use pathlib instead of os.path
    "PLW",   # pylint warnings
    "PLE",   # pylint errors
    "D",     # docstrings
]
ignore = [
    "D100",   # this warning cant be fixed with certain dependencies
    "E731",   # we need the lambda functionality so mypy doesn't error on sorted
    "E501",   # long lines after auto-formatting, such as long strings, are okay
    "ANN002", # we don't demand a type on *args, particularly because it's difficult to specify
    "ANN003", # we don't demand a type on **kwargs, particularly because it's difficult to specify
    "ANN401", # Allow "typing.Any"
    "N818",   # Don't require Exception class names to end with "Error"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "INP001", # Allow missing __init__.py files in tests
    "D", # Allow missing docstrings in tests
]

[tool.mypy]
strict = true
show_error_codes = true
strict_optional = false
disallow_any_generics = false  # allow ': dict' instead of ': dict[Any, Any]'
disable_error_code = "import-untyped"
exclude_gitignore = true

